+++ lib/src/shared/shared.dart
@@ -1,1 +1,2 @@
 export 'closable_box_controller.dart';
+export 'divider/sbb_divider.dart';