+++ CHANGELOG.md
@@ -13,6 +13,7 @@ It is expected that you keep this format strictly, since we depend on it in our 
 
 - added `SBBAccentButton`
 - added `SBBButtonSegment` for adding segments to `SBBSegmentedButton`
+- added `SBBDivider` for separating listed items within an indexed builder (e.g. `ListView.separated`)
 - added `SBBIllustration` for adding predefined illustrations to e.g. `SBBMessage`
 - added `SBBRadioGroup` - see [v5 migration guide]
 - added `SBBTextInput` as replacement for `SBBTextField` - see [v5 migration guide]